pyproject.toml
[tool.poetry]
name = "base-fastapi"
version = "0.1.0"
description = ""
authors = ["DinhQuang2312 <dinhquangcv231201@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.0"
alembic = "^1.8.1"
sqlmodel = "^0.0.8"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
uvicorn = {extras = ["uvloop"], version = "^0.23.1"}
psycopg2 = "^2.9.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
fastapi-mail = "1.2.0"
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
pre-commit = "^2.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
